refactor: move *.js files to /src
